requirements.txt
numpy==1.14.0
wordcloud==1.3.1
matplotlib==2.2.0
Pillow==5.0.0
